+++ source/renderer/assets/tpl/dialog/project-properties.handlebars.js
@@ -87,7 +87,13 @@ module.exports = {"1":function(container,depth0,helpers,partials,data) {
     + alias3(alias5(((stack1 = (depth0 != null ? depth0.pdf : depth0)) != null ? stack1.keywords : stack1), depth0))
     + "\" placeholder=\""
     + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.pdf.keywords",{"name":"i18n","hash":{},"data":data}))
-    + "\">\n        </div>\n        <div class=\"box-right\">\n          <label class=\"cb-toggle\">\n            <input type=\"checkbox\" value=\"yes\" name=\"pdf.titlepage\" id=\"pdf.titlepage\" "
+    + "\">\n          <hr>\n          <label for=\"textpl\">"
+    + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.pdf.textpl",{"name":"i18n","hash":{},"data":data}))
+    + "</label>\n          <div class=\"file-select-group\">\n            <input type=\"text\" name=\"pdf.textpl\" id=\"textpl\" value=\""
+    + alias3(alias5(((stack1 = (depth0 != null ? depth0.pdf : depth0)) != null ? stack1.textpl : stack1), depth0))
+    + "\">\n            <button type=\"button\" class=\"request-file\"\n            data-tippy-content=\""
+    + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.choose_file",{"name":"i18n","hash":{},"data":data}))
+    + "\"\n            data-request-name=\"TeX Template File\"\n            data-request-ext=\"tex\"\n            data-request-target=\"#textpl\"></button>\n          </div>\n        </div>\n        <div class=\"box-right\">\n          <label class=\"cb-toggle\">\n            <input type=\"checkbox\" value=\"yes\" name=\"pdf.titlepage\" id=\"pdf.titlepage\" "
     + ((stack1 = helpers["if"].call(alias1,((stack1 = (depth0 != null ? depth0.pdf : depth0)) != null ? stack1.titlepage : stack1),{"name":"if","hash":{},"fn":container.program(1, data, 0, blockParams, depths),"inverse":container.noop,"data":data})) != null ? stack1 : "")
     + ">\n            <div class=\"toggle\"></div>\n          </label>\n          <label for=\"pdf.titlepage\">"
     + alias3((helpers.i18n || (depth0 && depth0.i18n) || alias2).call(alias1,"dialog.preferences.pdf.titlepage_label",{"name":"i18n","hash":{},"data":data}))