+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Now only escaping characters are formatted, not the characters following them.
 - Fixed a bug that would prevent you from being able to modify an already loaded image without restarting Zettlr, because it would cache the image and not reload the modified version of it.
 - Updated the styling of form elements: Now ranges and radio buttons are also displayed in the Zettlr design.
+- Added an option to set a custom TeX template for PDF exports both in the general PDF preferences as well as on a per-project setting.
 
 ## Under the hood